Add discontinuation notice.
Thank you to everyone for using this modpack. As of Mar 26, 2025 the QuantuMiz EX is now officially merged into Main QuantuMiz modpack.

The reason of this decision: I had long concerns and questions, why it even needs to be split into different modpack while no one is even using Plasmo Voice?

Because of that I originally made up a plan to make QuantuMiz EX beyond what Vanilla can offer.

However some of popular mods that does it is closed-source or either incompatible with the FOSS license, which made it harder to implement 100% FOSS modpack that does the same.

So this is what happened. I have made decision to merge some of features in QMEX into original QM.

Without your support, I couldn't even get this far, having 3rd edition of QuantuMiz that aimed to provide improvement to the vanilla Minecraft.

The source code will remain as is so you are free to fork and make your own project. But please don't forget to change brandings and names to prevent confusion.

Thank you everyone. It was a long journey wasn't it?

It is time to say farewell to my third modpack. Thank you again for your love and support to the QuantuMiz project.